Changed the definition of generate_all_points because it was not working, corrected some mistakes on test.cpp

+++ include/epidemic1/SEIR.hpp
@@ -5,7 +5,7 @@ struct State {
     double E;
     double I;
     double R;
-    int N;
+    double N;
 };
 
 class sim {
